app/requirements/base.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-dynamic-model==0.3.0
psycopg2-binary==2.9.3
Pillow==9.1.0